Only register in router/search if item is live
This will break backwards compatibility but gives us needed flexibility.
The word 'live' is tricky - it's an attempt to distinguish this from publishing
workflow but may just add a new concept to keep track of. Would be good to get
some feedback on that.

A script will follow to mark all non-publisher artefacts 'live' and to make all
published publisher artefacts 'live'