Add an 'indexable_content' attribute to Artefact.
To pull out Rummager integration into observers, we should be using the
model instances to populate Rummager, but we don't want Panopticon to
have the responsibility for storing indexable content.

We have considered storing indexable content in Panopticon but it
doesn't feel right yet. This may be worth revisiting as the API approach
develops.