Upgrade to Rails 3.2.x (3.2.7)
Changes in artefacts_controller were to override the
new default behaviour which was making it return a
no content response and a 204 status. Relevant commit:
https://github.com/rails/rails/commit/aef62c4b4e89e7fa91973c188c9e7f281a44ecd0

Changes in integration_test_helper were because it
seemed to be getting into an infinite loop (although
the error was a SEGFAULT inside Rack).
Suspect it's from the Rack upgrade.