Take snapshots of all artefacts without actions.
To be able to see what someone has changed in an existing artefact, we
need an initial snapshot action to see what it was before. The simplest
way to do this is to go through all the artefacts and snapshot them.

This commit also fixes the UI to display the snapshots more helpfully,
because 'an unknown user did a "snapshot" action' isn't very helpful.