Add basic happy-path cucumber scenario for curated list csv upload
This gives us a very basic level of assurance that the happy path is
working. It should at some stage be supplemented with some sad-path
tests and a few other scenarios, but time is unlikely to allow that
for now. We should revisit this before we take this feature and make
it something we encourage content designers to use.