Use text instead of an icon for format
This lets us show something meaningful for Inside Government formats, without
needing to design icons just for this purpose, or making the column a mixture
of text and icons.

Speaking with Jon Sanger, the icons are useful in Publisher, as they allow
quick scanning of lists (eg "I can ignore all these because they're licences"),
but not so important in Panopticon.