Extend Artefact to fetch need information
Adds a method to the Artefact class called "need"
which, if the need owning service is Maslow, makes
a request to the Need API for information about the
need.

If any errors are raised by the API client, catch
these and return nil instead so that we don't break
the app.