Always allow the need_id for an Artefact to be edited
Once over, Need IDs that looked like Need-o-tron needs could not be
changed in Panopticon, for reasons unknown. When I added integration
with Maslow needs, I kept the same behaviour for Maslow needs.

We'd now like to change this so that any Need ID can be changed for an 
artefact. This will make it easier for content designers to keep the 
artefact up to date with the most appropriate need (and to more easily 
correct any mistakes).

I've removed the `need_id_editable?` method from the Artefact model 
extension class. I've also changed the form so that the Need ID field is 
always shown (and editable) regardless of whether the Need API returns 
information or not.