Include top-level tags in tag lists
There are use cases where things may need to be tagged with a top-level
tag in a tag type. To make this play nice within a multi-select field
and optgroups, this inserts the parent tag at the start of the optgroup.

Cherry pick Conflicts:

	test/integration/artefacts_edit_test.rb

I've not included the oil tags test, not needed by ODI.