Move to bootstrap-sass and Formtastic 2.2
This changes the app to use bootstrap-sass and updates our version of 
Formtastic to 2.2.1.

Bootstrap is now imported in application.scss and the previous overrides 
file is updated so that our custom styles still apply.

Using @import rather than manifest require statements is preferred so 
that we can extend Bootstrap styles in our own CSS rules.