Add an organisation importer
This adds a new class OrganisationImporter. It requests the list of all 
organisations from the API, and iterates over each result. If a tag 
doesn't already exist for the organisation, one will be created. If a 
tag already exists, then the tag title will be updated only if the 
organisation name has changed.

The importer is invoked through a new Rake task: organisations:import