Set up the Whenever gem
To keep the list of organisations up-to-date, we should run the importer 
on a regular basis. Our convention is to use the Whenever gem to set up 
app-based cron jobs.

The task definitions will live in alphagov-deployment, but this commit 
sets up Whenever, along with a placeholder config in schedule.rb.