Finish migration of EA mainstream sections and retagging
This finishes the work to add a migration for changing mainstream 
sections for the Environment Agency transition.

This migration:
- Creates a new parent section tag 'environment-countryside'
- Creates eight child section tags
- Retags mainstream content from the existing 'citizenship/
coasts-countryside' tag to the relevant child tags
- Adds additional tags to selected mainstream content
- Deletes the existing 'citizenship/coasts-countryside' tag

We are no longer assigning specialist sector tags as part of this. This 
work will now be done by the content team.

This depends on multiple other pieces of work:
- A migration in Whitehall to move detailed guide categories assigned to
the 'citizenship/coasts-countryside' tag (otherwise they will 500).
- Redirections in the Router for /browse/citizenship/coasts-countryside 
and the detailed guide categories.