Add curated list functionality for section tags
This takes the existing curated list functionality from the 'browse
section' controller, adding it to the new tags UI. I've added new
functional test coverage around this.

I've iterated this a bit, so that it doesn't increase the complexity
of the TagsController too much, and so it's styled to fit in with the
new tag form. However, it does come with a massive TODO noting that we
should improve how curated lists work soon.