Disable editing of related artefacts.
There are now c. 90k erteable tefacts.  This means that it's not
feasable to render these select inputs in their current form, and the
edit pages are currently timing out.

As a short-term measure, this disables these inputs so that the edit
form will work again.  A proper fix will follow as a separate piece of
work.