Updating relating items test for new ui
removed the integration test as we have
changed the functionality and all artefacts
are not present on the page beforehand, which
the test relies on. those features have been
covered in cucumber anyways.

removing the redirect assertion after destroying
relationship, because it doesn't work in tests
using poltergeist. redirects to publisher will
need mechanize to test. also, this bit gets
tested in many other tests.