Correct the slugs and type in the RPA migration
The migration was intended to remove some unused tags, but had an
incorrect tag type and an incorrect slug, so failed to work in
production.

This migration was run on staging, but not on production, and the
database has been synced from production to staging since then.
Therefore, we're just leaving the migration date as it is.