Allow draft sectors to be tagged to artefacts
This allows draft specialist sectors to be displayed in the
artefact edit form, so that they can be tagged to an artefact.

To make draft tags appear in the artefact edit form, we have to pass
an argument of `true` to the tag methods on `Artefact`.

To make this play nice with Rails' form helpers and conventions, I've
wrapped this behaviour in a decorator around `Artefact`.