Fix the JSON representation of the artefacts index.
Calling `respond_with` with more than one resource only really makes
sense in contexts where it's expected to create a redirect (eg in a
create action), and therefore calls `polymorphic_url`.  In this case it
was causing the json view of the index to render `@tag_collection`
instead of the artefacts.