Add a minimal JSON view for artefacts.
This view lets the URL arbiter <https://github.com/alphagov/url-arbiter>
seed its initial data set more quickly. Part of this is because the view
can skip tag lookups; part is simply that there is less JSON to
serialise. The response time on my dev VM dropped from 600-3000ms per
page to 50-100ms.

Tested together with the importer, limited to 20,000 of the 100,000 or
so artefacts, the time taken dropped from 6 minutes to 90 seconds.