Don't delete routes for Whitehall artefacts
This changes the UpdateRouterObserver so that it no longer deletes 
routes for artefacts which are owned by Whitehall.

Whitehall manages deletion of its own routes, and in many cases we 
actually wish to maintain the routes so that the custom 'unpublishing' 
behaviour continues to work.

However, we should continue to support registering Whitehall routes
as Whitehall now relies on Panopticon to register routes for detailed 
guides.