Support creating tags with JSON requests
This adds support for creating a tag in Panopticon by making a JSON
request. When a tag is created successfully, a 201 status is returned, 
along with a representation of the tag object. When a tag cannot be 
saved due to a validation error, we return a 422 status, along with 
details of the errors.

As actions already exist for HTML-formatted requests, the changes
here are mainly to support JSON requests on the same endpoints. There is
an additional change to allow tag attributes to be specified in requests
without the need for a root-level 'tag' entry.