Change tag routes to include the tag type and ID
So that other clients are able to make changes to tags without having to
know their internal BSON IDs, this changes Panopticon so that tag 
resources can be accessed at an endpoint which can be inferred from its
tag type and ID. This additional format looks like "/:tag_type/:tag_id", 
and supports tag_ids which contain slashes (eg. child tags). 

To make this change, I've changed config/routes.rb to add a route 
constraint to the tags resource so that slashes are permitted in the ID.
I've also added a new method to the Tag model (using the enhancements 
convention) which will find a Tag from its BSON ID or tag_id.