Stop prepending parent_id to tag_id on JSON create
This changes the TagsController so that we no longer prepend the 
parent_id of a tag to its tag_id when creating a child tag through a
JSON request.

This makes the behaviour of the JSON API more consistent with the 
update method, as the required value of tag_id will be the same across 
both the create and update methods.