Register email signup routes for topics.
Registers a prefix route for child topics and
deletes the exact route.

The copy to the array is because FOR SOME REASON
mongo/id returns the same item twice during an
each if you modify the item as it passes, with bad
data to boot.

WHY.  WHY.  WHY MONGO WHY.