Make specialist_sector_cleanup non-interactive
So that we can run the task easily using Fabric, we need to remove the
interactivity from the `specialist_sector_cleanup` Rake task.

The task performs only one operation, so the removal of interactivity is 
not a significant functionality regression.

I've also added some additional output in cases where the sector slug is
not provided.