Update specialist_sector artefact before registration
When a specialist sector is updated, we should update its artefact
before triggering a reindex of its content.  This avoids errors in the
reindexing causing the tag and artefact state to get out of sync (and
also means that if the reindex depended on the value of the artefact, we
wouldn't have a race condition).