Update poltergeist to work with latest phantomjs (v2.0)
I was seeing a bunch of tests failing with the following error:

    Could not find an executable 'phantomjs' that matched the requirements '~> 1.8', '>= 1.8.1'. Found versions were {"/usr/local/bin/phantomjs"=>"2.0.0"}.

The error appears to be related to an incompatibility between poltergeist
v1.3.0 and phantomjs v2.0 which has been fixed in this issue [1] and released
in poltergeist v1.6.0 [2].

The newer version of poltergeist seems to work ok with pre-2.0 versions of
phantomjs or at least it does with phantomjs v1.9.8.

c.f. https://github.com/alphagov/calculators/pull/112

[1]: https://github.com/teampoltergeist/poltergeist/issues/566
[2]: https://github.com/teampoltergeist/poltergeist/blob/master/CHANGELOG.md#160