Allow PUT of a tag to create instead of 404.
Collections-publisher is the canonical source of tags. It created them
here via the API when they're created there. It also updates them on any
changes. Currently, if the create_tag call errors for any reason, any
further updates will also fail with a 404, meaning that it's not
possible to get the tag into panopticon without devops intervention.

This updates the tag PUT action to create the tag if it doesn't already
exist, which avoids the problem when creation fails.