Avoid Artefact.any_instance.stubs(:id) in tests.
There's something odd going on between Mocha and Mongoid when on ruby
2.1 that meant that after these tests had run, the id method no longer
existed on Artefact, causing all following tests to fail. Commenting out
these stubs allowed all the other tests to pass.

I haven't tracked down the exact cause of the problem.  Upgrading to the
latest Mocha didn't help. Given the fact that this app is being
deprecated, I've taken the decision to work around the problem as
opposed to spending more time getting to the root of it.