Use assert_hash_including
Exactly comparing the return value of
`RummagableArtefact#artefact_hash` with our expectations means we can't
easily add keys to the output.

`assert_hash_including` verifies that our *expected* hash is "included
in" the resulting hash. This means we don't have to change all the
tests when adding something to the output.

This works in the same way as the rspec include matcher works.

https://www.relishapp.com/rspec/rspec-expectations/v/2-0/docs/matchers/i
nclude-matcher#hash-usage