Only update editions when the artefact slug changes
Before this change, any time an Artefact was updated in Panopticon, the edition
would have `save!` called on it, which would update the `updated_at` field on
all the non-archived editions. For example, this included hundreds of items
where the content was re-tagged to live in the new Topics.

This was a problem because the updated_at field is used to see what editions
have been worked on recently - it is really important in the content designers'
workflow.

Given that the intent was only to update the slug on the editions if it had
changed on the artefact then let's only call the method when the slug has
changed.

The alternative would be to update the slug on the editions without triggering
callbacks, but that seems potentially risky and expectation-subverting.