Add migration to fix duplicate artefacts
During the Detailed Guide migration, there was a window where the slug
that Whitehall used to publish artefacts had been updated to be
prefixed with “guidance/“, but Panopticon’s database hadn’t been
migrated to result the artefacts. This meant that when editors
published changes to existing Detailed Guides, Whitehall published a
new artefact with the new slug.

This migration deletes the newer artefact (in order to preserve the
original ID), merges any changes that were made during the publish and
then performs the reslug as required. It also appends the change
history from the duplicated artefact to the original, but changing
“create” actions to “update”