Ensure correct paths are set for Detailed Guide artefacts
This loops through all Detailed Guide artefacts and:
- Ensures the slug begins with `guidance/`, and isn’t prefixed with
`deleted-`
- Ensures the paths are correct, including translations. This will
publish the correct routes to the Router API. We don’t need the
`/:slug` form any more, and Panopticon won’t delete routes from the
router as a result of removing a path from the paths array. We need to
do this to ensure the root redirect isn’t overwritten when published.
We check first that the root form is owned by Whitehall in order to
prevent publishing `/guidance/:slug` when `/:slug` is already a
redirect away to another content type.