Avoid mongo cursor timeout in query
Mongo fetches records for this query in batches of 1000. The script
takes over 10 minutes to publish each batch on preview, which means the
cursor has timed out by the time it goes to fetch the next batch. To
work around this, we fetch them all in advance and iterate over the
array. This fits in under 400MB of RAM and avoids the timeout issue