Remove faking of section in travel-advice search results
We are sending a "fake" mainstream browse section to rummager in this
application.

This was added to display the correct breadcrumb in the search result
metadata. We no longer display the that information in the metadata
(removed in https://github.com/alphagov/frontend/pull/806), so we don't
need this anymore.

Because rummager copies the section information to the
`mainstream_browse_pages` field, this will remove the
`mainstream_browse_pages` tag `foreign-travel-advice` from the rummager
documents.

This is okay because the `foreign-travel-advice` section is not an
existing mainstream browse page:

  https://www.gov.uk/browse/foreign-travel-advice

However, filtering by `mainstream_browse_pages` will no longer be
possible. But there's no evidence that this is ever used.

https://www.gov.uk/api/search.json?filter_mainstream_browse_pages[]=fore
ign-travel-advice