Only allow manual artefact creation for publisher formats
Currently it's possible to create artefacts with any format that
doesn't belong to `whitehall` or `panopticon `. This is odd because the
`owning_app` is set to `publisher` by default for these docs, via a
hidden field in the form.

This behaviour causes non-publisher artefacts created this way to show
"This content is managed in Publisher." with a link that will show an
error.

We need to be strict about formats and owning apps because we wish to
give all documents `content_ids`, and those attributes will tell us who
the should generate that content_id.

To verify we're not breaking existing functionality, there is a rake
task that prints out per-app the number of documents that don't have
the default `owning_app`. This currently shows no relevant mismatching
artefacts.