Don't run edition validations when archiving an artefact
The bump in the gem gets this fix.

Before this change, attempting to archive an artefact with invalid editions
would successfully mark the artefact as archived, but the editions would be
left in their state.

We have published editions which are invalid because validation rules have been
introduced since they were last worked on/saved.

Ideally, the state machine transition for archiving would skip validations, but
state_machine doesn't seem to support this.

Successfully tested with known-problem examples in Panopticon in development.