Bump gds-sso to pin transitive dependencies
Currently, a `bundle update` on this repo will update the
`omniauth-oauth2` gem (a dependency of gds-sso) to 1.4.0. This version
is not compatible with signon.

To prevent accidental breakage we have pinned to version in `gds-sso`.
This commit bumps the gem to that version so that we can run `bundle
update` safely again for panopticon.

PRs on gds-sso:
- https://github.com/alphagov/gds-sso/pull/84
- https://github.com/alphagov/gds-sso/pull/83