Save organisation content_id in Tag sync
The content_id of organisation isn't saved in the tag currently.

Rails has dirty tracking on models and will not send data to the
database that hasn't changed. The manual checking therefore is not
necessary.

By making it less dependent on the title we can expand the attributes
to be copied later, like `content_id`.

The content_id we need to start migrating taggings to the
publishing-api.