Prevent tags being entered in panopticon GUI for content owned by migrated apps
Panopticon currently allows tags to be entered for all artefacts.  In future, we want to prevent tags being entered in panopticon for content owned by apps which have been migrated to the new tag manager app.  This is implemented by the use of a config file which specifies which apps are to have the tagging option suppressed (config/settings/<environment>.yml.

If 'publisher' is included in this list, then tagging is suppressed for new content entered in panopticon, as all new content is given this owning_app by default, otherwise it just relates to editing existing content.