Set parent link as primary section
We are splitting up the mainstream_browse_page into `parent` (used for
breadcrumb) and the main tags.

This commit makes sure that if we send a `parent` link, we copy that
tag to the front of the section tags. This will ensure that the same
browse page will be displayed in the breadcrumb on mainstream browse
pages.