Add ManualArtefactGuarantor
This object will act to make sure that a given manual has a corresponding artefact in panopticon to allow for tagging.  We need this because specialist-publisher currently doesn't create these and it's a) quite complicated to change b) being rewritten anway.

For now this object only checks that the supplied slug actually resolves to an item in the content store.