Use old id format in JSON representation.
Mongoid now outputs a model's id in JSON as a nested hash with an $oid
key. Other apps such as Publisher have been upgraded to use this new
format, but using it here would break everything that uses this API. So
we patch the as_json method to convert back to the old format.