Update list of non migrated apps
This commit does several things:

- Removes config gem
- Removes config initializer that reads the settings.yml file
- Removes settings.yml file
- uses NON_MIGRATED_APPS constant in Artefact object instead

All this changes aim to ease future changes, since migrating apps to the
new tagging is a work in progress at the moment.

This NON_MIGRATED_APPS constant should be in sync with
[Content
Tagger](https://github.com/alphagov/content-tagger)
[blacklisted-apps](https://github.com/alphagov/content-tagger/blob/master/config/blacklisted-apps.yml)

And with [Rummanger](https://github.com/alphagov/rummager)
[NON_MIGRATED_APPS](https://github.com/alphagov/rummager/blob/master/lib/indexer/index_documents.rb#L6)
constant.

By removing the config gem we were able to see that  TaggingUpdater was
stil coupled with the settings.yml file, if we had not removed it this
would had reached production with a silent bug that was not catched by
the test suite.

Part of:
https://trello.com/c/tZd23gW5/568-allow-tagging-in-content-tagger-for-apps-that-can-t-be-tagged-yet-s