Do not allow organisation tagging for specialist-content
It is currently possible to add organisation tags to
specialist-publisher owned content. This is a mistake, because
specialist-publisher also sends organisations.

If you would tag a document to an organisation here, it would
eventually end up in rummager (because rummager reads tags from the
content-api). For clarity, organisations should only be set by
specialist-publisher.