Parse 'sections' param correctly to fix tagging of mainstream browse pages
When param handling was converted to use strong parameters there were
a number of oversights. This one broke mainstream browse page tagging
from mainstream publisher.

It's not clear whether the strong parameters feature is at fault -
introducing it seems to have caused several bugs, but testing in this area was
rather patchy.

@benhyland and @tijmenb