This test wasn't thoroughly covering the case
before:
- There was an Artefact without links.
- A message without links about that Artefact was being processed.
The only thing tested was whether the message was being acked.
The effect on the Artefact wasn't being tested, hence this test was not useful at all.

now:
- There is an Artefact with a link.
- A message with no links about the Artefact is being processed.
The number of links of the Artefact is being tested to check whether the message has had any effect on it.