Fix bug in sending content to rummager
This is a hotfix for a bug that appears in panopticon when it gets
called from panopticon.

Panopticon is sending a `indexable_content` parameter, which propagates
through the controller to something that decides whether to send the
document to rummager as a update or create:

https://github.com/alphagov/panopticon/blame/adc087a30a2e4f5013b32c0cb3f
2569a9d7b3696/app/models/rummageable_artefact.rb#L56

Because `indexable_content` is not wrapped, it will be nil and triggers
the `should_amend` code path. This will fail in rummager because the
document doesn’t exist.

A follow up to this will fix the root causes of this problem.