Move scopes into class instead of mixin
The method to extract code into modules: provides a false sense that
you have “thin models”, even though your public interface stays as big
as it is.

See
http://blog.codeclimate.com/blog/2012/10/17/7-ways-to-decompose-fat-acti
verecord-models/ for some thinking on this.

It’s better to inline the definition here so that readers know what
methods are exposed on the model.