Allow tags to be cleared by tagging updater
The TaggingUpdater listens to the message queue and copies things from
the `links` hash to this app's database. This is necessary to allow
tags set in content-tagger to propagate to the frontend.

This commit allows us to clear out the links for an item. Because
publishing-api never sends empty link sets this previously didn't
happen.