Make the find-your-local-council artefact a prefix route
This is to support the new interactive find-your-local-council being
built in frontend that replaces the local directgov version of the tool.

We update the find-your-local-council artefact to swap the
/find-your-local-council path from living in paths (this generates an
exact route in router) to living in prefixes (which generates a prefix
route in router). Saving the artefact sends the relevant messages to
router-api to propagate the change up through the stack.