Restrict external links to publisher and smart-answer pages
We are planning on moving the external links to another app so that
panopticon can be retired. Currently this feature is used in 149
publisher-owned pages, and 1 smart-answers page.

A current plan is to move the functionality over to mainstream
publisher and come up with a different solution for the one
smart-answer.

By disabling it for all other applications we make sure that we don't
have to implement this feature in other applications as well.