Don't show content-tagger link on "new" page
Currently the "new artefact" page has a link to content-tagger. Because
new artefacts don't have a content_id yet, the link goes nowhere and
teh user will see a 404.

This adds a check if there's a `content_id` for the artefact. It also
allows us to remove old code which had the same purpose (content by
finder-api, frontend and planner didn't have a content_id).