Remove rummager dependency from panopticon
Previously, panopticon was sending artifacts to Rummager. That flow is now being
deprecated in favour of a simpler approach which is to send documents directly
to rummager.

The following applications have been updated in order to send documents directly
into Rummager:

- [x] Calendars
- [x] Calculators
- [x] Business Support Finder
- [x] Licence Finder
- [x] Smart Answers
- [ ] Publisher
- [ ] Travel Advice Publisher

When all those applications start running that rake task after deploy, we can
merge this PR.

Trello ticket: https://trello.com/c/gsHnT5WF/161-epic-send-things-to-rummager-directly-instead-of-via-panopticon