Add search admin warning to withdraw screen
We're in the process of removing the Rummager dependency from
Panopticon. Once this is done, Panopticon will no longer tell Rummager
to delete withdrawn artefacts from search results. Editors will need to
do this manually in Search Admin. This commit adds some warning text to
the withdraw screen to remind the editor to do this.