Add functionality to remove an artefact from search on withdraw
In 54a927e we removed Rummager as a dependency in Panopticon. We need to
add one piece of functionality back in - the deletion of content from
the search index when an artefact is withdrawn.

This commit adds a service object to handle this, invoking it from
ArtefactsController#destroy. Because the call is synchronous, the
service object implements some basic error handling. It tries contacting
Rummager twice and then reports to Airbrake in the event of any further
errors, leaving the destroy action to complete transparently for the
user.