Only allow artefact withdrawal for publisher-artefacts
It's currently possible for editors to "withdraw" artefacts owned by
all applications except Whitehall in this app. We think that this is
could cause trouble. Things like manuals and travel advice have their
own withdrawal behaviour and interact with the publishing-api, unlike
this application.

To prevent potential problems, we restrict withdrawal to artefacts
owned by publisher. In the near future this feature will be moved to
publisher outright.