Send related links to publishing-api
This commit makes sure that we send the related links of an item to the
publishing-api every time it is saved. This will allow us to start
rendering the sidebar in frontend apps from the content-store.

A follow up PR will create a rake task to republish all existing
artefacts.