Lock down route registration for artefacts
Currently, artefacts owned by any app except `whitehall` can have their
routes registered from this app.

This is dangerous because there 1) is a lot of old content in this
database, 2) some of the routes for the old artefacts have since been
registered "properly" via the publishing-api, with different routes.

There's a chance that panopticon will start sending the routes for one
of these artefacts, overriding the correct routes set by the
publishing-api.

This locks down the artefacts for which we'll register routes. The six
are:

-business-support-finder
- calculators
- calendars
- license-finder
- publisher
- smart-answers

These are the only apps that currently rely on panopticon for route
registration.