Remove tags from parameter parsing
This class determines which parameters can be sent to the controller
(via API and the panopticon UI).

These tags will only ever be updated via the message queue, never via
this controller. This should have been done when we switched over to
use the message queue as the only way of modifying tags here, but this
was forgotten.