Remove organisations tagging related code
Back story: it used to be possible to tag artefacts to organisations in
the Panopticon UI. This information was forwarded to search, but not
used in the `frontend` application for display to users.

Since we're now indexing the organisations in search directly from the
publishing-api, it's no longer necessary to retain the organisations in
panopticon.

This commit removes the functionality to import the organisations from
whitehall (the "organisations API"). It then removes some code related
to changing URLs of organisations, and the code that makes the message
queue sync copy organisation taggings into this shared database.