Fix error with rake routes
This fixes the error identified in [mainstream migration discovery][1]
where running `rake routes` results in the following error:

RegexpError: end pattern with unmatched parenthesis: /^\/artefacts\/((?x-mi:[^\.]+#1ormorenon-dotcharacters(\.[a-z]{2}#optionallydotfollowedbyabasiclocale(eg'fr')(#optionallyfollowedbyalocaleextension-[a-z]{2}|#egzh-hk-\d{3}#eges-419)?)?))\/history(?:\.([^\/.?]+))?$/
/usr/lib/rbenv/versions/2.2.3/bin/bundle:23:in `load'
/usr/lib/rbenv/versions/2.2.3/bin/bundle:23:in `<main>'

It looks like [free spacing mode][2] isn't working properly anymore for
`rake routes`.  It works OK in Rails 3.2 which was when the constraint was
added.

I found two options that worked:
- Moving the comments to outside the regex (what I did)
- Using the non-freespace comment style inline:
  - (?# This is a comment)

I chose not to use the inline comment style as it looks slightly like it might
still be code.

I've also wrapped the styleguide in a conditional as there is no need for this
to be present in a production environment.

[1]: https://trello.com/c/WvBbjT89/87-panopticon-s-rake-routes-errors-with-a-regex-unmatched-parenthesis-error
[2]: https://ruby-doc.org/core-2.1.1/Regexp.html#class-Regexp-label-Free-Spacing+Mode+and+Comments